Fixing issue BerandaViewModel and Create Menu Layanan

+++ app/src/main/java/com/stp/maunyucibeta/model/beranda/Menu.kt
@@ -2,7 +2,7 @@ package com.stp.maunyucibeta.model.beranda
 
 data class Menu(
     var id: Int = 0,
-    var omset: Double = 0.0,
+    var omset: Int = 0,
     var masuk: Int = 0,
     var ambil: Int = 0,
     var selesai: Int = 0
